requirements.txt
asgiref==3.8.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==4.0.4
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-phonenumber-field==7.3.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
google-api-core==2.19.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mailersend==0.5.6
packaging==24.1
phonenumbers==8.13.37
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
rsa==4.9
sqlparse==0.5.0
stripe==9.7.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
